small refactor, concurrency, auth bool
